ensure committee meeting protocols rewrite all ids to descriptor, even if not downloaded

+++ datapackage_pipelines_knesset/committees/processors/parse_committee_meeting_protocols.py
@@ -26,20 +26,22 @@ class ParseCommitteeMeetingProtocolsProcessor(BaseProcessor):
         text_relpath = os.path.join(str(meeting_protocol["committee_id"]), "{}.txt".format(meeting_protocol["meeting_id"]))
         parts_filename = self._get_filename(parts_relpath)
         text_filename = self._get_filename(text_relpath)
+        protocol_filename = meeting_protocol["protocol_file"]
         if parts_relpath not in self._all_filenames:
             self._all_filenames += [parts_relpath, text_relpath]
             os.makedirs(os.path.dirname(parts_filename), exist_ok=True)
-        with CommitteeMeetingProtocol.get_from_filename(meeting_protocol["protocol_file"]) as protocol:
-            with open(text_filename, "w") as f:
-                f.write(protocol.text)
-            with open(parts_filename, "w") as f:
-                csv_writer = csv.writer(f)
-                csv_writer.writerow(["header", "body"])
-                for part in protocol.parts:
-                    csv_writer.writerow([part.header, part.body])
+        if not os.path.exists(parts_filename):
+            with CommitteeMeetingProtocol.get_from_filename(protocol_filename) as protocol:
+                with open(text_filename, "w") as f:
+                    f.write(protocol.text)
+                with open(parts_filename, "w") as f:
+                    csv_writer = csv.writer(f)
+                    csv_writer.writerow(["header", "body"])
+                    for part in protocol.parts:
+                        csv_writer.writerow([part.header, part.body])
         yield {"committee_id": meeting_protocol["committee_id"],
                "meeting_id": meeting_protocol["meeting_id"],
-               "protocol_file": meeting_protocol["protocol_file"],
+               "protocol_file": protocol_filename,
                "text_file": text_filename,
                "parts_file": parts_filename}
